backend/requirements.txt
# Run these libraries before accessing model
flask==2.3.3
flask-cors==4.0.0
werkzeug==2.3.7
groq
SpeechRecognition
transformers
textblob
termcolor
python-dotenv
torch
torchvision
torchaudio
tiktoken
protobuf
happytransformer
gradio>=4.0.0
spacy
nltk
librosa
numpy
pandas
matplotlib
seaborn
plotly
scipy
kaleido  # For saving plotly figures as images
pydub  # For audio format conversion
psutil  # For process management